third_party/blink/web_tests/svg/dynamic-updates/SVGFEImageElement-dom-preserveAspectRatio-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'preserveAspectRatio' attribute of the SVGFEImageElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS image.getAttribute('preserveAspectRatio') is "xMinYMin slice"
PASS image.getAttribute('preserveAspectRatio') is "xMinYMid slice"
PASS successfullyParsed is true

TEST COMPLETE
